refactor(web): move workspace_statuses reads into infra (#535 follow-up 7)
Resolve the parked TODO(#313 follow-up) in ProjectService.list.

- New infra/db/queries/workspace-statuses.ts owns the read-side SQL
  for the workspace_statuses table: loadCurrent() + getByWorkspaceId().
- services/state.ts::{loadCurrentStatuses, getWorkspaceStatus} now
  delegate to the new query class (back-compat surface preserved for
  unrelated callers).
- ProjectService takes a WorkspaceStatusQueries dependency via its
  constructor, calls it directly, and references WorkspaceAgentInfo
  from infra/events/status-event-bus instead of the
  ReturnType<typeof loadCurrentStatuses> trick. TODO block removed.

Refs: https://github.com/band-app/band/issues/535
